requirements.txt
pykka
RPi.GPIO
paho-mqtt
